Move ed25519-donna.h to ed25519-donna subfolder
