+++ src/Solana/Program.cpp
@@ -10,7 +10,7 @@
 #include "../Base58.h"
 #include "../Hash.h"
 
-#include "TrezorCrypto/ed25519-donna.h"
+#include <TrezorCrypto/ed25519-donna/ed25519-donna.h>
 
 #include <cassert>
