+++ trezor-crypto/crypto/ed25519-donna/curve25519-donna-scalarmult-base.c
@@ -1,4 +1,4 @@
-#include <TrezorCrypto/ed25519-donna.h>
+#include <TrezorCrypto/ed25519-donna/ed25519-donna.h>
 #include <TrezorCrypto/ed25519.h>
 
 /* Calculates nQ where Q is the x-coordinate of a point on the curve
