+++ trezor-crypto/crypto/ed25519-donna/ed25519-donna-basepoint-table.c
@@ -1,4 +1,4 @@
-#include <TrezorCrypto/ed25519-donna.h>
+#include <TrezorCrypto/ed25519-donna/ed25519-donna.h>
 
 /* multiples of the base point in packed {ysubx, xaddy, t2d} form */
 const uint8_t ALIGN(16) ge25519_niels_base_multiples[256][96] = {
