+++ trezor-crypto/crypto/tests/test_check.c
@@ -47,7 +47,7 @@
 #include <TrezorCrypto/blake2s.h>
 #include <TrezorCrypto/curves.h>
 #include <TrezorCrypto/ecdsa.h>
-#include <TrezorCrypto/ed25519-donna.h>
+#include <TrezorCrypto/ed25519-donna/ed25519-donna.h>
 #include <TrezorCrypto/ed25519-donna/ed25519-keccak.h>
 #include <TrezorCrypto/ed25519.h>
 #include <TrezorCrypto/memzero.h>
